requirements.txt
# Core
PyYAML
pandas
torch
transformers
accelerate

# PDF
PyPDF2

# Testing
pytest

